+++ src/index.ts
@@ -1,5 +1,5 @@
 import { GoogleCalendarEventFetcher } from "./event-fetcher.js";
 
 export * from "./event-fetcher.js";
-export type * from "./google-calendar.js";
+export * from "./google-calendar.js";
 export default GoogleCalendarEventFetcher;